Draw embedded images only a single time
Embedded images were previously drawn once for each character
in the embed key, causing issues for semi-transparent images.
See issue #278. With this commit, the embed image is only drawn
for the first character in the embed keys, and a sentinel
attribute is used for the remaining characters in the key to
draw neither text not an image for them.